requirements.txt
pandas>=1.5.0
scikit-learn>=1.2.0
supabase>=2.0.0
python-dotenv>=1.0.0
